Moved feature config from "patches"
