+++ A3A/addons/ultimate/CfgEventHandlers.hpp
@@ -1,0 +1,42 @@
+class Extended_InitPost_EventHandlers
+{
+    class A3AU_TerrainSmoother_Base_F 
+    {
+        class ADDON 
+        {
+            clientInit = QUOTE(call A3A_fnc_handlerTerrainObjectHiderPostInitClient);
+        };
+    };
+
+    class A3AU_VegetationCleaner_Base_F 
+    {
+        class ADDON 
+        {
+            clientInit = QUOTE(call A3A_fnc_handlerTerrainObjectHiderPostInitClient);
+        };
+    };
+
+    class GVAR(BB_TerrainObjectHider_Base) 
+    {
+        class ADDON 
+        {
+            clientInit = QUOTE(call A3A_fnc_handlerTerrainObjectHiderPostInitClient);
+        };
+    };
+};
+
+class Extended_PostInit_EventHandlers 
+{
+    class ADDON
+	{
+        init = "call A3U_fnc_postInit";
+    };
+};
+
+class Extended_PreInit_EventHandlers 
+{
+    class ADDON
+	{
+        init = "call A3U_fnc_init";
+    };
+};
