+++ A3A/addons/ultimate/CfgSounds.hpp
@@ -1,0 +1,6 @@
+class GVAR(LockpickToolBreak)
+{
+    name = "Lockpick tool breaking";
+    sound[] = {"x\A3A\addons\ultimate\data\lockpick-break.ogg", 1, 1, 25};
+    titles[] = {};
+};
