+++ A3A/addons/ultimate/config.cpp
@@ -19,9 +19,8 @@ class CfgPatches
 
 class CBA_Extended_EventHandlers_base; // Needed for CfgVehicles.hpp
 
+#include "CfgEventHandlers.hpp"
 #include "CfgFunctions.hpp"
-#include "patches\CfgMagazines.hpp"
-#include "patches\CfgSounds.hpp"
 
 class A3U
 {
@@ -30,50 +29,17 @@ class A3U
     #include "config\trader\cfgTraderMods.hpp"
 };
 
+class CfgMagazines
+{
+    #include "CfgMagazines.hpp"
+};
+
+class CfgSounds
+{
+    #include "CfgSounds.hpp"
+};
+
 class CfgVehicles
 {
     #include "CfgVehicles.hpp"
-};
-
-class Extended_InitPost_EventHandlers
-{
-    class A3AU_TerrainSmoother_Base_F 
-	{
-        class ADDON 
-    {
-            clientInit = QUOTE(call A3A_fnc_handlerTerrainObjectHiderPostInitClient);
-    };
-    };
-
-    class A3AU_VegetationCleaner_Base_F 
-    {
-        class ADDON 
-    {
-            clientInit = QUOTE(call A3A_fnc_handlerTerrainObjectHiderPostInitClient);
-    };
-    };
-
-    class GVAR(BB_TerrainObjectHider_Base) 
-    {
-        class ADDON 
-        {
-            clientInit = QUOTE(call A3A_fnc_handlerTerrainObjectHiderPostInitClient);
-    };
-};
-};
-
-class Extended_PostInit_EventHandlers 
-{
-    class ADDON
-	{
-        init = "call A3U_fnc_postInit";
-    };
-};
-
-class Extended_PreInit_EventHandlers 
-{
-    class ADDON
-	{
-        init = "call A3U_fnc_init";
-    };
 };
