--- A3A/addons/ultimate/patches/CfgSounds.hpp
@@ -1,7 +1,0 @@
-class CfgSounds {
-    class GVAR(LockpickToolBreak) {
-        name = "Lockpick tool breaking";
-        sound[] = {"x\A3A\addons\ultimate\data\lockpick-break.ogg", 1, 1, 25};
-        titles[] = {};
-    };
-};
